Introduce an example component without @computed
ComponentB is more or less the same as ComponetA but it doesn't rely on
the use of `@computed` in a subclass of `@glimmer/component` (which isn't
a subclass of `@ember/object`.

We still see the same problems with this component (e.g. not tracking nested
properties and blowing up with:

> trap returned extra keys but proxy target is non-extensible

when using the latest `ember-source` canary or beta.